third_party/WebKit/LayoutTests/platform/mac/fast/forms/select/select-clientheight-large-size-expected.txt

Tests that select elements cap their size to the size attribute and to 4 when no size is specified.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS clientHeight('select2') is clientHeight('select1')
PASS getElemById('select2').setAttribute('size', '4'); clientHeight('select2') == clientHeight('select1') is true
PASS getElemById('select2').setAttribute('size', '5'); clientHeight('select2') > clientHeight('select1') is true
PASS getElemById('select2').setAttribute('size', '8'); clientHeight('select2') is within 1 of 114
PASS getElemById('select2').setAttribute('size', '12'); clientHeight('select2') is within 1 of 171
FAIL getElemById('select2').setAttribute('size', '16'); clientHeight('select2') should be within 1 of 224. Was 227.
PASS successfullyParsed is true

TEST COMPLETE
